item_search/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-multipart==0.0.9
odfpy==1.4.1
python-docx==1.1.2
openpyxl==3.1.5
pytesseract==0.3.13
pdf2image==1.17.0
Pillow==10.4.0
numpy==1.26.4

